backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai==1.12.0
tiktoken==0.5.2
requests==2.31.0
